feat: decouple project structure
