+++ schemas.py
@@ -1,0 +1,10 @@
+from flask_marshmallow import Marshmallow
+
+ma = Marshmallow()
+
+class NoteSchema(ma.Schema):
+    class Meta:
+        fields = ("id", "title", "content")
+
+note_schema = NoteSchema()
+notes_schema = NoteSchema(many=True)
